Added more features to model, question about Gama
I have been trying to ensure that agents created initially in the model
are transfered to residential spaces. I did that by creating a list of
residential spaces and using the one_of operator to ensure that the
agent assigned to a random residential space, but I for some reason my
commands only populate my list of residential spaces as "nil" this
occurs in lines 37-46.