Improved Model 12/31/16
This model provides a more complete implementation of the paper we went over when we first started the project.

The model entirely follows the rules outlined by the powerpoint I had made a few weeks ago. In addition there are a number of charts also displayed so various metrics of the model can be tracked as it cycles.

The next step I feel is running experiments on the model so one can observe "interesting" results. Mostly what this model does is just give into the tipping point of a few creative spaces spawning the entire grid becoming creative. In addition, the distribution of landuse ("residential spaces", "green spaces", etc.) is entirely random so it would be interesting to see what would happen with different designed land use distributions.

Thanks and have a wonderful new years,
Ken